+++ R/frmtmb-package.R
@@ -1,7 +1,7 @@
 #' @keywords internal
-#' @importFrom stats anova coef confint cooks.distance deviance
-#'   df.residual extractAIC family fitted formula influence logLik
-#'   model.frame model.matrix nobs predict profile residuals sigma
-#'   simulate terms update vcov weights
+#' @importFrom stats anova coef confint cooks.distance deviance dfbeta
+#'   dfbetas df.residual drop1 extractAIC family fitted formula
+#'   influence logLik model.frame model.matrix na.action nobs predict
+#'   profile residuals sigma simulate terms update vcov weights
 #' @importFrom graphics plot
 "_PACKAGE"